+++ account_e-invoice_generate/README.rst
@@ -1,24 +1,32 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
    :alt: License: AGPL-3
 
 =============================
 Electronic Invoice Generation
 =============================
 
-This is a technical module needed to ensure compatibility between the two electronic invoice generation modules: *account_invoice_ubl* and *account_invoice_factur-x*. These 2 modules are able to embed an XML file in the PDF invoice, but for the moment only one XML file can be embeded in the PDF. And it can be useful to have both modules installed, because, for example, you may need to generate Factur-X PDF invoices for some customers and UBL XML files for other customers. So it adds a configuration parameter to decide which XML format is embedded in the PDF.
+This is a technical module needed to ensure compatibility between
+the two electronic invoice generation modules: *account_invoice_ubl*
+and *account_invoice_factur-x*. These 2 modules are able to embed an
+XML file in the PDF invoice, but for the moment only one XML file can
+be embedded in the PDF. And it can be useful to have both modules installed,
+because, for example, you may need to generate Factur-X PDF invoices
+for some customers and UBL XML files for other customers. So it adds
+a configuration parameter to decide which XML format is embedded in the PDF.
 
 Configuration
 =============
 
-The new configuration parameter *XML Format embedded in PDF invoice* is available in the menu *Accounting > Configuration > Settings*.
+The new configuration parameter *XML Format embedded in PDF invoice*
+is available in the menu *Accounting > Configuration > Settings*.
 
 Usage
 =====
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/226/10.0
+   :target: https://runbot.odoo-community.org/runbot/226/11.0
 
 Bug Tracker
 ===========
@@ -26,7 +34,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues
 <https://github.com/OCA/edi/issues>`_. In case of trouble, please
 check there if your issue has already been reported. If you spotted it first,
-help us smashing it by providing a detailed and welcomed feedback.
+help us smash it by providing detailed and welcomed feedback.
 
 Credits
 =======
@@ -35,6 +43,7 @@ Contributors
 ------------
 
 * Alexis de Lattre <alexis.delattre@akretion.com>
+* Andrea Stirpe <a.stirpe@onestein.nl>
 
 Maintainer
 ----------
