+++ account_e-invoice_generate/__init__.py
@@ -1,3 +1,3 @@
-# -*- coding: utf-8 -*-
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import models
