+++ account_e-invoice_generate/__manifest__.py
@@ -1,19 +1,19 @@
-# -*- coding: utf-8 -*-
-# © 2018 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# Copyright 2018 Akretion (http://www.akretion.com)
+# @author: Alexis de Lattre <alexis.delattre@akretion.com>
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 {
     'name': 'Account e-invoice Generate',
-    'version': '10.0.1.0.0',
+    'version': '11.0.1.0.0',
     'category': 'Accounting & Finance',
     'license': 'AGPL-3',
     'summary': 'Technical module to generate PDF invoices with '
                'embedded XML file',
     'author': 'Akretion,Odoo Community Association (OCA)',
-    'website': 'http://www.akretion.com',
+    'website': 'https://github.com/OCA/edi',
     'depends': ['account'],
     'data': [
-        'views/account_config_settings.xml',
+        'views/res_config_settings.xml',
     ],
     'installable': True,
 }
