+++ account_e-invoice_generate/models/__init__.py
@@ -1,4 +1,4 @@
-# -*- coding: utf-8 -*-
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from . import company
-from . import account_config_settings
+from . import res_company
+from . import res_config_settings
