+++ account_e-invoice_generate/models/res_config_settings.py
@@ -1,0 +1,12 @@
+# Copyright 2018 Akretion (http://www.akretion.com)
+# @author: Alexis de Lattre <alexis.delattre@akretion.com>
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class ResConfigSettings(models.TransientModel):
+    _inherit = 'res.config.settings'
+
+    xml_format_in_pdf_invoice = fields.Selection(
+        related='company_id.xml_format_in_pdf_invoice')
