+++ account_e-invoice_generate/readme/CONFIGURE.rst
@@ -1,0 +1,2 @@
+The new configuration parameter *XML Format embedded in PDF invoice*
+is available in the menu *Accounting > Configuration > Settings*.
