+++ account_e-invoice_generate/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Alexis de Lattre <alexis.delattre@akretion.com>
+* Andrea Stirpe <a.stirpe@onestein.nl>
