+++ account_e-invoice_generate/views/res_config_settings.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  Copyright 2018 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+  License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+<record id="view_account_config_settings" model="ir.ui.view">
+    <field name="model">res.config.settings</field>
+    <field name="inherit_id" ref="account.res_config_settings_view_form"/>
+    <field name="arch" type="xml">
+        <xpath expr="//div[@id='invoicing_settings']" position="after">
+            <h2>Electronic Invoices</h2>
+            <div class="row mt16 o_settings_container" id="invoicing_settings">
+                <div class="col-xs-12 col-md-6 o_setting_box" id="e-invoices">
+                    <div class="o_setting_left_pane"/>
+                    <div class="o_setting_right_pane">
+                        <label string="Select Format"/>
+                        <span class="fa fa-lg fa-building-o" title="Values set here are company-specific." groups="base.group_multi_company"/>
+                        <div class="text-muted">
+                            Select the format of your e-invoices.
+                        </div>
+                        <div class="content-group" name="xml_format_in_pdf_invoice">
+                            <div class="row mt16">
+                                <label string="Format" class="col-md-3 o_light_label"/>
+                                <field name="xml_format_in_pdf_invoice"/>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </xpath>
+    </field>
+</record>
+
+</odoo>
